Added feature set path to group_vars and host_vars (#101)
* Added feature set path to group_vars and host_vars

Added feature to allow setting the directory path where the group_vars
and host_vars are located. This is an optional feature and as such it
does not break the existing fuctionality. This would allow the
adjustment of the search paths to support multi-staged environment
layouts such as the Alternative Direcory Layout at
http://docs.ansible.com/ansible/playbooks_best_practices.html

* Added test cases

* Fixed syntax error in tests

* Updated to use separate test case folders

* Moved roles dir to playbook base directory

* Removed group1.yml from groups_all tests