Updated to fix variable resolution issue with parent child groups
The current code base does away with [parent:children] groups in favor
of converting the parent group into a group at the same level its
children. However, this dismisses the fact that parent child
relationships are a means to add a level of precedence to variable
assignments across groups. This commit adds this logic to better align
ansible_specs variable assignments with how this is done in Ansible.